also share StateExploredEventMessage, i.e. move it to converter bundle.
