+++ bundles/org.palladiosimulator.analyzer.slingshot.stateexploration/src/org/palladiosimulator/analyzer/slingshot/stateexploration/explorer/DefaultGraphExplorer.java
@@ -17,6 +17,7 @@ import org.eclipse.emf.ecore.util.EcoreUtil;
 import org.palladiosimulator.analyzer.slingshot.common.events.DESEvent;
 import org.palladiosimulator.analyzer.slingshot.converter.StateGraphConverter;
 import org.palladiosimulator.analyzer.slingshot.converter.data.StateGraphNode;
+import org.palladiosimulator.analyzer.slingshot.converter.events.StateExploredEventMessage;
 import org.palladiosimulator.analyzer.slingshot.core.Slingshot;
 import org.palladiosimulator.analyzer.slingshot.core.api.SimulationDriver;
 import org.palladiosimulator.analyzer.slingshot.core.api.SystemDriver;
@@ -31,7 +32,6 @@ import org.palladiosimulator.analyzer.slingshot.stateexploration.explorer.config
 import org.palladiosimulator.analyzer.slingshot.stateexploration.explorer.configuration.UriBasedArchitectureConfiguration;
 import org.palladiosimulator.analyzer.slingshot.stateexploration.explorer.planning.Postprocessor;
 import org.palladiosimulator.analyzer.slingshot.stateexploration.explorer.planning.Preprocessor;
-import org.palladiosimulator.analyzer.slingshot.stateexploration.messages.StateExploredEventMessage;
 import org.palladiosimulator.analyzer.slingshot.stateexploration.providers.AdditionalConfigurationModule;
 import org.palladiosimulator.analyzer.slingshot.stateexploration.rawgraph.DefaultGraph;
 import org.palladiosimulator.analyzer.slingshot.stateexploration.rawgraph.DefaultState;
@@ -104,7 +104,7 @@ public class DefaultGraphExplorer implements GraphExplorer {
 		this.fringe = new FringeFringe(new PriorityTransitionQueue()); // new FIFOTransitionQueue()
 
 		systemDriver.postEvent(
-				new StateExploredEventMessage(this.convertState(this.graph.getRoot(), null, null)));
+				new StateExploredEventMessage(this.convertState(this.graph.getRoot(), null, null), "Explorer"));
 
 		this.preprocessor = new Preprocessor(this.graph, this.fringe,
 				LaunchconfigAccess.getMinDuration(launchConfigurationParams));
@@ -214,7 +214,7 @@ public class DefaultGraphExplorer implements GraphExplorer {
 
 		current.setUtility(value);
 
-		this.systemDriver.postEvent(new StateExploredEventMessage(node));
+		this.systemDriver.postEvent(new StateExploredEventMessage(node, "Explorer"));
 
 		// TODO : this is temporal. remove later on. Actually this is a reasonable idea
 		// to include for the prioritazion of the fringe.
